requirements.txt
streamlit>=1.30.0
pandas
pdfplumber
PyPDF2
reportlab
google-generativeai
